Print the Content-Length header only if it actually exists
It is unsettling to see the "Content-Length" header printed in the diagnosis messages even though it is not actually present in the request and/or response headers.

Note that we can't just test whether the `ContentLength` property is null or not because the [HttpContentHeaders.ContentLength implementation][1] automatically computes a value even if the Content-Length is not present. This is why using `content.Headers.TryGetValues` is mandatory.

[1]: https://github.com/dotnet/runtime/blob/v8.0.5/src/libraries/System.Net.Http/src/System/Net/Http/Headers/HttpContentHeaders.cs#L41-L55